requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
evdev==1.9.1
idna==3.10
pynput==1.8.1
pyperclip==1.9.0
python-xlib==0.33
requests==2.32.3
six==1.17.0
urllib3==2.4.0
dotenv
tenacity
